Discover and check operator variants
Operator variants can now be explicitly specified in the OpInfo kwargs.
When the operator name is not the same as the method/function form this
will allow them to be discovered.

The OpInfo is extended to also accept/discover the inplace operator
variant.

Operator and inplace operator variants are exercised in consistency
tests when the sample does not contain any kwargs.

ghstack-source-id: bdc988349a60bbb7512215627c9e620bc0b0d673
Pull Request resolved: https://github.com/pytorch/pytorch/pull/76695