[LT] Add a trie data structure for caching IR nodes
Summary: TrieCache provides a way to look up an IR node before we
actually create it. If the lookup hits in TrieCache, we reuse the
existing node and move the current pointer in TrieCache to point to that
node; if the lookup misses, we create a new node and insert it into TrieCache.

ghstack-source-id: d71d7af43c3273278bc769d6081e6ff57cbcb473
Pull Request resolved: https://github.com/pytorch/pytorch/pull/76542